move socless_template_string to integrations.py
